requirement.txt
fastapi
uvicorn
python-jose[cryptography]
python-dotenv
pydantic
groq